week_13/day1/ThemeParkStartPoint/src/test/java/people/VisitorTest.java
package people;

import attractions.Attraction;
import attractions.RollerCoaster;
import org.junit.Before;
import org.junit.Test;

import javax.swing.*;

import static org.junit.Assert.assertEquals;

public class VisitorTest {

    Visitor visitor;
    RollerCoaster rollerCoaster;

    @Before
    public void before(){
        visitor = new Visitor(14, 1.2, 40.0);
    }

    @Test
    public void hasAge() {
        assertEquals(14, visitor.getAge());
    }

    @Test
    public void hasHeight() {
        assertEquals(1.2, visitor.getHeight(), 0.1);
    }

    @Test
    public void hasMoney() {
        assertEquals(40.0, visitor.getMoney(), 0.1);
    }

    @Test
    public void hasVisited2Attractions(){
        rollerCoaster = new RollerCoaster("Tower", 2);
        visitor.addAsVisited(rollerCoaster);
        visitor.addAsVisited(rollerCoaster);
        assertEquals(2, visitor.howManyVisited());

    }
}
